Remove more dead code
