+++ app/src/main/report/components/Artefacts/ArtefactsList.tsx
@@ -1,6 +1,5 @@
 import * as React from "react";
 import {ILookup} from "../../../shared/models/Lookup";
-import {settings} from "../../../shared/Settings";
 import {Artefact} from "../../../shared/models/reports/Artefact";
 import {ArtefactItem} from "./ArtefactItem";
