--- app/src/main/shared/models/RemoteContent.ts
@@ -1,3 +1,0 @@
-export interface RemoteContent {
-    ready: boolean;
-}
