+++ app/src/test/contrib/components/Responsibilities/Overview/ResponsibilityOverviewPageTests.tsx
@@ -5,7 +5,7 @@ import { Store } from "redux";
 
 import "../../../../helper";
 import { Sandbox } from "../../../../Sandbox";
-import {createMockContribStore, createMockStore} from "../../../../mocks/mockStore";
+import {createMockContribStore} from "../../../../mocks/mockStore";
 import {PageArticle} from "../../../../../main/shared/components/PageWithHeader/PageArticle";
 import {mockMatch} from "../../../../mocks/mocks";
 import {
