+++ app/src/test/mocks/mockResult.ts
@@ -1,0 +1,15 @@
+import { ErrorInfo, Result, ResultStatus } from "../../main/shared/models/Generated";
+
+export function mockResult<T>(data: T,
+                              errors?: Array<ErrorInfo>,
+                              status?: ResultStatus): Result {
+    errors = errors || [];
+    status = status || "success";
+
+    return { data, errors, status };
+}
+
+export function mockError(code: string, message: string): ErrorInfo {
+    return { code, message };
+}
+
