+++ app/src/test/report/components/Reports/ReportDownloadsTests.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import {shallow} from "enzyme";
 
 import {expect} from "chai";
-import {mockReport, mockVersion} from "../../../mocks/mockModels";
+import {mockVersion} from "../../../mocks/mockModels";
 import {Sandbox} from "../../../Sandbox";
 import {ArtefactsSection} from "../../../../main/report/components/Artefacts/ArtefactsSection";
 import {DataLinks} from "../../../../main/report/components/Data/DataLinks";
