+++ app/src/test/report/components/Resources/ResourceLinksTest.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { mount, shallow } from "enzyme";
+import { shallow } from "enzyme";
 import {expect} from "chai";
 import {ResourceLinks} from "../../../../main/report/components/Resources/ResourceLinks";
 import { FileDownloadLink } from "../../../../main/shared/components/FileDownloadLink";
